add security to users

+++ api/controllers/usuario-permiso.controller.js
@@ -24,11 +24,30 @@ class UsuarioPermisoController {
     }
 
     async createUsuarioPermiso(req, res) {
-        const { body } = req;
-        const createdUsuarioPermiso = await this._usuarioPermisoService.create(body);
-        return res.status(201).send({
-            payload: createdUsuarioPermiso
-        });
+
+        let token = req.headers.authorization.split(' ')[1];
+        let payload = jwt.verify(token, 'secretKey');
+        const idUser = payload.subject;
+        const listaPermisos = await this._usuarioPermisoService.getAll();
+        let permitido = false;
+        for (let i = 0; i < listaPermisos.length; i++) {
+            const permiso = listaPermisos[i];
+
+            if (idUser == permiso.UsuarioId && permiso.PermisoId == 2) {
+                permitido = true;
+            }
+
+        }
+
+        if (permitido) {
+            const { body } = req;
+            const createdUsuarioPermiso = await this._usuarioPermisoService.create(body);
+            return res.status(201).send({
+                payload: createdUsuarioPermiso
+            });
+        } else {
+            return res.status(401).send();
+        }
     }
 
     async updateUsuarioPermiso(req, res) {
@@ -40,10 +59,29 @@ class UsuarioPermisoController {
     }
 
     async deleteUsuarioPermiso(req, res) {
-        const { id } = req.params;
 
-        await this._usuarioPermisoService.delete(id);
-        return res.status(204).send();
+        let token = req.headers.authorization.split(' ')[1];
+        let payload = jwt.verify(token, 'secretKey');
+        const idUser = payload.subject;
+        const listaPermisos = await this._usuarioPermisoService.getAll();
+        let permitido = false;
+        for (let i = 0; i < listaPermisos.length; i++) {
+            const permiso = listaPermisos[i];
+
+            if (idUser == permiso.UsuarioId && permiso.PermisoId == 2) {
+                permitido = true;
+            }
+
+        }
+
+        if (permitido) {
+            const { id } = req.params;
+            await this._usuarioPermisoService.delete(id);
+            return res.status(204).send();
+        } else {
+            return res.status(401).send();
+        }
+
     }
 }
